Add final score based on player deaths

+++ Assets/Scripts/GameManager.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 // attach to a GO in the first scene only (switched it to the storyline scene)
 public class GameManager : MonoBehaviour
@@ -9,8 +10,16 @@ public class GameManager : MonoBehaviour
     {
         if (isFirstScene)
         {
-            PlayerPrefs.DeleteAll();
-            PlayerPrefs.Save();
+            if (SceneManager.GetActiveScene().name == "ending")
+            {
+                PlayerScoreTracker.ClearPrefsButKeepFinalScore();
+            }
+            else
+            {
+                PlayerPrefs.DeleteAll();
+                PlayerScoreTracker.ResetScore();
+            }
+
             Debug.Log("[GameManager] Fresh game - stats wiped.");
         }
     }
